+++ libtracepoint/src/tracepoint.c
@@ -133,12 +133,11 @@ user_events_data_update(int* staticFileOrError)
     - Can be made to work in a container.
     */
 
-    // Start by probing the most likely absolute paths.
+    // Start by probing the most likely absolute path.
     // This improves performance and also makes it simpler to set up a container:
     // The container creator will have to project the user_events_data file but
     // won't need to make a dummy /proc/mounts file for us to parse.
-    if (0 <= (newFileOrError = open("/sys/kernel/tracing/user_events_data", O_RDWR)) ||
-        0 <= (newFileOrError = open("/sys/kernel/debug/tracing/user_events_data", O_RDWR)))
+    if (0 <= (newFileOrError = open("/sys/kernel/tracing/user_events_data", O_RDWR)))
     {
         // Success.
     }
@@ -205,32 +204,14 @@ user_events_data_update(int* staticFileOrError)
                 continue;
             }
 
-            char const* path_suffix;
-            size_t path_suffix_len; // includes NUL
-            char keepLooking;
+            char const* const path_suffix = "/user_events_data";
+            size_t const path_suffix_len = sizeof("/user_events_data"); // includes NUL
 
             const char* const pchTracefs = "tracefs";
             size_t const cchTracefs = sizeof("tracefs") - 1;
-            const char* const pchDebugfs = "debugfs";
-            size_t const cchDebugfs = sizeof("debugfs") - 1;
 
             size_t const fs_len = fs_end - fs_begin;
-            if (fs_len == cchTracefs && 0 == memcmp(line + fs_begin, pchTracefs, cchTracefs))
-            {
-                // "tracefsMountPoint/user_events_data"
-                path_suffix = "/user_events_data";
-                path_suffix_len = sizeof("/user_events_data"); // includes NUL
-                keepLooking = 0; // prefer "tracefs" over "debugfs".
-            }
-            else if (path[0] == 0 &&
-                fs_len == cchDebugfs && 0 == memcmp(line + fs_begin, pchDebugfs, cchDebugfs))
-            {
-                // "debugfsMountPoint/tracing/user_events_data"
-                path_suffix = "/tracing/user_events_data";
-                path_suffix_len = sizeof("/tracing/user_events_data"); // includes NUL
-                keepLooking = 1; // prefer "tracefs" over "debugfs".
-            }
-            else
+            if (fs_len != cchTracefs || 0 != memcmp(line + fs_begin, pchTracefs, cchTracefs))
             {
                 continue;
             }
@@ -246,23 +227,19 @@ user_events_data_update(int* staticFileOrError)
             memcpy(path, line + mount_begin, mount_len);
             memcpy(path + mount_len, path_suffix, path_suffix_len); // includes NUL
 
-            if (!keepLooking)
-            {
-                break;
-            }
+            break;
         }
 
         fclose(mountsFile);
 
         if (path[0] == 0)
         {
-            // No "tracefs" or "debugfs" mount point found.
+            // No "tracefs" mount point found.
             newFileOrError = -ENOTSUP;
         }
         else
         {
-            // path is now something like "/sys/kernel/tracing/user_events_data\0" or
-            // "/sys/kernel/debug/tracing/user_events_data\0".
+            // path is now something like "/sys/kernel/tracing/user_events_data\0".
             newFileOrError = open(path, O_RDWR);
             if (0 > newFileOrError)
             {